+++ src/pages/ReviewReplyPage/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useHandleReceiver } from './useHandleReceiver'
